pyproject.toml
[project]
name = "medical-learning-assistant"
version = "0.1.0"
description = "Multi-Agent Medical Learning Assistant with RAG"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    
    # AI/ML
    "groq>=0.4.1",
    "google-generativeai>=0.3.2",
    "sentence-transformers>=2.3.1",
    
    # Vector Database
    "pinecone>=5.0.0",  # Updated from pinecone-client
    
    # MySQL
    "mysql-connector-python==9.2.0",
    "pymysql>=1.1.0",
    
    # Web Search & Scraping
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.1.0",
    
    # PDF Processing
    "pypdf2>=3.0.1",
    "pdfplumber>=0.10.3",
    
    # Utilities
    "python-dotenv>=1.0.1",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false
dev-dependencies = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
]